+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 - Classes & structs that conform to SwiftUI's PreviewProvider are now retained.
 - Support @main entry points.
 - Fix unused recursive function detection when using the index store.
+- Properties named in struct implicit constructors are now retained.
 
 ## 1.8.0 (2020-10-3)